The `SharedBuf` read/write now return `usize` rather than `SysCallReg`
Also changes some `u64` to `usize` to be consistent with the `Read` and `Write`
traits.

+++ src/main/bindings/c/bindings.h
@@ -156,7 +156,7 @@ struct ByteQueue *bytequeue_new(uintptr_t default_chunk_size);
 
 void bytequeue_free(struct ByteQueue *bq_ptr);
 
-uint64_t bytequeue_numBytes(struct ByteQueue *bq);
+uintptr_t bytequeue_numBytes(struct ByteQueue *bq);
 
 bool bytequeue_hasBytes(struct ByteQueue *bq);
